fpl/src/requirements.txt
# python version = 3.10
kedro~=0.18.8
kedro-datasets~=1.2.0
scikit-learn~=1.2.2
hyperopt~=0.2.7
flatten_dict~=0.4.2
deepdiff~=6.3.0
xgboost~=1.7.5
ipykernel~=6.15.0
seaborn~=0.12.2
matplotlib~=3.7.1
pulp~=2.7.0
cryptography~=40.0.2
paramiko~=3.1.0
regex~=2023.5.5
pyyaml~=6.0
pycaret~=3.0.2